requirements.txt
flask==1.0.2
numpy==1.15.1
matplotlib==2.2.3
pandas==0.23.4
kombu==4.2.1
nltk==3.3
scikit_learn==0.19.2
scipy==1.1.0
cassandra-driver==3.14.0
requests==2.19.1
gunicorn==19.9.0
aiohttp==3.4.4
aiohttp-wsgi==0.8.2